deleting not used files
