+++ contracts/marketplace/src/testing/helpers/nft_functions.rs
@@ -71,7 +71,7 @@ pub fn transfer(
     let res = router.execute_contract(creator.clone(), collection.clone(), &transfer_msg, &[]);
     assert!(res.is_ok());
 }
-#[allow(dead_code)]
+
 pub fn burn(router: &mut StargazeApp, creator: &Addr, collection: &Addr, token_id: u32) {
     let transfer_msg: Sg721ExecuteMsg<Empty, Empty> = Sg721ExecuteMsg::Burn {
         token_id: token_id.to_string(),
