+++ contracts/marketplace/src/testing/setup.rs
@@ -1,9 +1,6 @@
 pub mod constants;
 pub mod msg;
 pub mod setup_accounts;
-pub mod setup_second_bidder_account;
-// pub mod setup_contracts;
-pub mod setup_marketplace;
-// pub mod setup_minter;
 pub mod setup_contracts;
+pub mod setup_marketplace;
 pub mod templates;
